requirements.txt
Django~=5.1.7
psycopg2~=2.9.10
python-dotenv~=1.1.0
Pillow~=11.1.0
channels~=4.2.2
channels-redis~=4.2.1
daphne~=4.1.2
gunicorn~=23.0.0
uvicorn~=0.33.0
ruff~=0.11.10
asgiref~=3.8.1
whitenoise~=6.9.0
drf-spectacular~=0.28.0
djangorestframework~=3.16.0